requirements.txt
streamlit
requests
python-dotenv
pandas
scikit-learn
gdown
# any other packages you use
